Add @KotlinOpen to class annotations only
    The kotlin-open is making the class and the members as well public,
    so adding this annotation to member AA annotation is not needed has
    no effect.

+++ dataset/GitHub_Java/androidannotations.androidannotations/AndroidAnnotations/androidannotations-core/androidannotations-api/src/main/java/org/androidannotations/annotations/Background.java
@@ -1,5 +1,6 @@
 /**
- * Copyright (C) 2010-2015 eBusiness Information, Excilys Group
+ * Copyright (C) 2010-2016 eBusiness Information, Excilys Group
+ * Copyright (C) 2016-2018 the AndroidAnnotations project
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -191,7 +192,7 @@ public @interface Background {
 	 *
 	 * @return the delay of the execution
 	 */
-	int delay() default 0;
+	long delay() default 0;
 
 	/**
 	 * Serial execution group.
